rename old files to have a uniformed format
